Posterior Probability
The task in this part is to implement a system that:
1) Can determine the posterior probability of different hypotheses,
given priors for these hypotheses, and given a sequence of observations.
2) Can determine the probability that the next observation will be of a
specific type, priors for different hypotheses, and given a sequence of
observations.

The five possible hypotheses for our bag are:
h1 (prior: 10%): This type of bag contains 100% cherry candies.
h2 (prior: 20%): This type of bag contains 75% cherry candies and 25%
lime candies.
h3 (prior: 40%): This type of bag contains 50% cherry candies and 50%
lime candies.
h4 (prior: 20%): This type of bag contains 25% cherry candies and 75%
lime candies.
h5 (prior: 10%): This type of bag contains 100% lime candies.